fix to deleting plant

+++ backend/database_layer.go
@@ -207,17 +207,24 @@ func (handler *DatabaseHandler) CreateNewSchedule(
 }
 
 func (handler *DatabaseHandler) DeletePlant(user_id string, plant_id int) (string, error) {
-	query := `
-		DELETE FROM plants WHERE user_id = $1 AND plant_id = $2;
-		DELETE FROM schedule WHERE plant_id = $2;
-    `
-
-	result, err := handler.Db.Exec(query, user_id, plant_id)
+	tx, err := handler.Db.Begin()
 	if err != nil {
-		return "", fmt.Errorf("failed to delete plant: %v", err)
+		return "", fmt.Errorf("failed to start transaction: %v", err)
 	}
 
-	rowsAffected, err := result.RowsAffected()
+	defer tx.Rollback()
+
+	res1, err := tx.Exec("DELETE FROM plants WHERE user_id = $1 AND plant_id = $2", user_id, plant_id)
+	if err != nil {
+		return "", fmt.Errorf("failed to delete from plants: %v", err)
+	}
+
+	_, err = tx.Exec("DELETE FROM schedule WHERE plant_id = $2", plant_id)
+	if err != nil {
+		return "", fmt.Errorf("failed to delete from schedule: %v", err)
+	}
+
+	rowsAffected, err := res1.RowsAffected()
 	if err != nil {
 		return "", fmt.Errorf("unable to check rows affected: %v", err)
 	}
@@ -226,5 +233,6 @@ func (handler *DatabaseHandler) DeletePlant(user_id string, plant_id int) (strin
 		return "", fmt.Errorf("no plant found for given user and plant_id")
 	}
 
+	tx.Commit()
 	return "Plant deleted successfully", nil
 }
